refactor(auth): reorganize module structure
