+++ app/src/androidTest/java/timur/gilfanov/messenger/application/NavigationApplicationTest.kt
@@ -24,7 +24,6 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import timur.gilfanov.messenger.MainActivity
 import timur.gilfanov.messenger.annotations.ApplicationTest
 import timur.gilfanov.messenger.data.repository.DefaultIdentityRepository
 import timur.gilfanov.messenger.data.repository.LocaleRepositoryImpl
@@ -44,6 +43,7 @@ import timur.gilfanov.messenger.test.AndroidTestDataHelper.DataScenario.NON_EMPT
 import timur.gilfanov.messenger.test.AndroidTestRepositoryWithRealImplementation
 import timur.gilfanov.messenger.test.RepositoryCleanupRule
 import timur.gilfanov.messenger.test.SettingsRepositoryStub
+import timur.gilfanov.messenger.ui.activity.MainActivity
 import timur.gilfanov.messenger.util.Logger
 
 @OptIn(ExperimentalTestApi::class)
