+++ app/src/main/AndroidManifest.xml
@@ -20,7 +20,7 @@
         android:theme="@style/Theme.Messenger"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.activity.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.Messenger">
             <intent-filter>
