+++ feature/auth/src/androidTest/java/timur/gilfanov/messenger/feature/auth/LoginFeatureTest.kt
@@ -32,14 +32,14 @@ import timur.gilfanov.messenger.annotations.FeatureTest
 import timur.gilfanov.messenger.auth.R
 import timur.gilfanov.messenger.auth.di.AuthModule
 import timur.gilfanov.messenger.auth.di.AuthViewModelModule
-import timur.gilfanov.messenger.auth.login.GoogleSignInClient
-import timur.gilfanov.messenger.auth.login.GoogleSignInClientStub
-import timur.gilfanov.messenger.auth.login.GoogleSignInResult
-import timur.gilfanov.messenger.auth.login.LoginScreenTestActivity
-import timur.gilfanov.messenger.auth.login.LoginWithCredentialsUseCase
-import timur.gilfanov.messenger.auth.login.LoginWithCredentialsUseCaseImpl
-import timur.gilfanov.messenger.auth.login.LoginWithGoogleUseCase
-import timur.gilfanov.messenger.auth.login.LoginWithGoogleUseCaseImpl
+import timur.gilfanov.messenger.auth.domain.usecase.LoginWithCredentialsUseCase
+import timur.gilfanov.messenger.auth.domain.usecase.LoginWithCredentialsUseCaseImpl
+import timur.gilfanov.messenger.auth.domain.usecase.LoginWithGoogleUseCase
+import timur.gilfanov.messenger.auth.domain.usecase.LoginWithGoogleUseCaseImpl
+import timur.gilfanov.messenger.auth.ui.GoogleSignInClient
+import timur.gilfanov.messenger.auth.ui.GoogleSignInClientStub
+import timur.gilfanov.messenger.auth.ui.GoogleSignInResult
+import timur.gilfanov.messenger.auth.ui.LoginScreenTestActivity
 import timur.gilfanov.messenger.auth.validation.CredentialsValidatorImpl
 import timur.gilfanov.messenger.domain.entity.ResultWithError
 import timur.gilfanov.messenger.domain.entity.auth.AuthState.Unauthenticated
