+++ feature/auth/src/debug/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application>
-        <activity android:name="timur.gilfanov.messenger.auth.login.LoginScreenTestActivity" />
+        <activity android:name="timur.gilfanov.messenger.auth.ui.LoginScreenTestActivity" />
     </application>
 </manifest>
