+++ feature/auth/src/debug/kotlin/timur/gilfanov/messenger/auth/ui/LoginScreenTestActivity.kt
@@ -1,4 +1,4 @@
-package timur.gilfanov.messenger.auth.login
+package timur.gilfanov.messenger.auth.ui
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -8,6 +8,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.ui.Modifier
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
+import timur.gilfanov.messenger.auth.ui.screen.login.LoginScreen
 import timur.gilfanov.messenger.ui.theme.MessengerTheme
 
 @AndroidEntryPoint
