+++ feature/auth/src/main/kotlin/timur/gilfanov/messenger/auth/di/AuthModule.kt
@@ -8,10 +8,10 @@ import javax.inject.Singleton
 import kotlinx.coroutines.CoroutineScope
 import timur.gilfanov.messenger.auth.AuthInterceptor
 import timur.gilfanov.messenger.auth.AuthTokenStorage
-import timur.gilfanov.messenger.auth.TokenRefreshError
-import timur.gilfanov.messenger.auth.TokenRefreshUseCase
-import timur.gilfanov.messenger.auth.login.GoogleSignInClient
-import timur.gilfanov.messenger.auth.login.GoogleSignInClientImpl
+import timur.gilfanov.messenger.auth.domain.usecase.TokenRefreshError
+import timur.gilfanov.messenger.auth.domain.usecase.TokenRefreshUseCase
+import timur.gilfanov.messenger.auth.ui.GoogleSignInClient
+import timur.gilfanov.messenger.auth.ui.GoogleSignInClientImpl
 import timur.gilfanov.messenger.auth.validation.CredentialsValidatorImpl
 import timur.gilfanov.messenger.domain.entity.ResultWithError
 import timur.gilfanov.messenger.domain.entity.auth.AuthProvider
