+++ feature/auth/src/main/kotlin/timur/gilfanov/messenger/auth/domain/usecase/LoginWithGoogleUseCase.kt
@@ -1,4 +1,4 @@
-package timur.gilfanov.messenger.auth.login
+package timur.gilfanov.messenger.auth.domain.usecase
 
 import timur.gilfanov.messenger.domain.entity.ResultWithError
 import timur.gilfanov.messenger.domain.entity.auth.GoogleIdToken
