+++ feature/auth/src/main/kotlin/timur/gilfanov/messenger/auth/domain/usecase/TokenRefreshUseCaseImpl.kt
@@ -1,4 +1,4 @@
-package timur.gilfanov.messenger.auth
+package timur.gilfanov.messenger.auth.domain.usecase
 
 import timur.gilfanov.messenger.domain.entity.ResultWithError
 import timur.gilfanov.messenger.domain.entity.auth.AuthTokens
