+++ feature/auth/src/main/kotlin/timur/gilfanov/messenger/auth/ui/GoogleSignInClient.kt
@@ -1,4 +1,4 @@
-package timur.gilfanov.messenger.auth.login
+package timur.gilfanov.messenger.auth.ui
 
 import android.content.Context
