+++ feature/auth/src/test/kotlin/timur/gilfanov/messenger/auth/ui/LoginViewModelProcessDeathTest.kt
@@ -1,4 +1,4 @@
-package timur.gilfanov.messenger.auth.login
+package timur.gilfanov.messenger.auth.ui
 
 import androidx.lifecycle.SavedStateHandle
 import app.cash.turbine.test
@@ -9,7 +9,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.experimental.categories.Category
 import timur.gilfanov.messenger.annotations.Component
-import timur.gilfanov.messenger.auth.login.LoginViewModelTestFixtures.createViewModel
+import timur.gilfanov.messenger.auth.ui.LoginViewModelTestFixtures.createViewModel
 import timur.gilfanov.messenger.testutil.MainDispatcherRule
 
 @OptIn(ExperimentalCoroutinesApi::class)
