+++ feature/auth/src/testFixtures/kotlin/timur/gilfanov/messenger/auth/ui/GoogleSignInClientStub.kt
@@ -1,4 +1,4 @@
-package timur.gilfanov.messenger.auth.login
+package timur.gilfanov.messenger.auth.ui
 
 import android.content.Context
 import kotlinx.coroutines.CompletableDeferred
