+++ feature/auth/src/testFixtures/kotlin/timur/gilfanov/messenger/auth/ui/LoginViewModelTestFixtures.kt
@@ -1,6 +1,9 @@
-package timur.gilfanov.messenger.auth.login
+package timur.gilfanov.messenger.auth.ui
 
 import androidx.lifecycle.SavedStateHandle
+import timur.gilfanov.messenger.auth.domain.usecase.LoginWithCredentialsUseCaseImpl
+import timur.gilfanov.messenger.auth.domain.usecase.LoginWithGoogleUseCaseImpl
+import timur.gilfanov.messenger.auth.ui.screen.login.LoginViewModel
 import timur.gilfanov.messenger.domain.entity.ResultWithError
 import timur.gilfanov.messenger.domain.entity.auth.AuthSession
 import timur.gilfanov.messenger.domain.entity.auth.validation.CredentialsValidationError
